sql: move tableInserter to its own file
Release note: None
